% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_associate.R
\name{tds_associate}
\alias{tds_associate}
\title{A function to combine the text with a labelled set of metadata.}
\usage{
tds_associate(dir, training_metadata = "", csv_type = "comma",
  max_char = 5e+05)
}
\arguments{
\item{dir}{A directory containing the training corpora, with one document per text file.}

\item{training_metadata}{The name of the file containing the name of the name of the documents, the labels and possibly other metadata.}

\item{csv_type}{A type of csv file: "comma" (by default) uses read_csv while "semicolon" uses read_csv2.}

\item{max_char}{The maximum number of characters per documents (500000 by default). In practive, it should not be a problem.}
}
\value{
A data frame combining the texts located in a directory and the labelled data in a tsv/csv file.
}
\description{
A function to combine the text with a labelled set of metadata.
}
\examples{
 tds_associate(dir = "corpora", training_metadata = "my_corpora.csv")
}
