% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_matrix.R
\name{tds_matrix}
\alias{tds_matrix}
\title{A function to transform tf_idf counts into a matrix ready for svm classification}
\usage{
tds_matrix(text_count)
}
\arguments{
\item{text_count}{A tidy dataframe of textual data pre-processed by tds_process}
}
\value{
A processed version of the textual dataset using tf_idf values with cosine normalization.
}
\description{
A function to transform tf_idf counts into a matrix ready for svm classification
}
\examples{
 tds_matrix(text_count)
}
