% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_model.R
\name{tds_model}
\alias{tds_model}
\title{A function to train an SVM model.}
\usage{
tds_model(text_count, classification_matrix = FALSE, prob_state = TRUE,
  min_label = 0, max_label = 0, cost_variable = 0)
}
\arguments{
\item{text_count}{A tidy dataframe of textual data pre-processed by tds_process}

\item{classification_matrix}{A term/document matrix. Set to false (by default) to create the matrix from text_count.}

\item{prob_state}{Set to true (default) to get the detailed probabilities. This is highly recommended but can take longer.}

\item{min_label}{The minimal number of document per labels. Default is no threshold (0). If some labels are poorly represented and harder to predict it can be useful to drop them.}

\item{max_label}{The maximal number of document per labels. Default is no threshold (0). SVM is sensible to over-fitting: if a label is too represented in the original corpora, it will over-populate the results.}

\item{cost_variable}{A parametrized value for svm classification (purely optional)}
}
\value{
An SVM model that can be interpreted using tds_open.
}
\description{
A function to train an SVM model.
}
\examples{
 tds_model(text_count, classification_matrix = FALSE, prob_state = TRUE, min_label = 70, max_label = 150, cost_variable = 0)
}
