% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_open.R
\name{tds_open}
\alias{tds_open}
\title{A function to output the word probabilities inside a model.}
\usage{
tds_open(model, show_graphic = TRUE, display_token = 10)
}
\arguments{
\item{model}{An SVM model.}

\item{show_graphic}{If set to TRUE (default), using the function will produce a graphic of the most important word per labels.}

\item{display_token}{The number of word per labels to show in the graphics.}
}
\value{
A tidy dataset from the predictions
}
\description{
A function to output the word probabilities inside a model.
}
\examples{
 tds_retrieve(predictions)
}
