% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_retrieve.R
\name{tds_retrieve}
\alias{tds_retrieve}
\title{A function to get the predictions result in a tidy format.}
\usage{
tds_retrieve(predictions)
}
\arguments{
\item{predictions}{The predictions generated by SVM.}
}
\value{
A tidy dataset from the predictions
}
\description{
A function to get the predictions result in a tidy format.
}
\examples{
 tds_retrieve(predictions)
}
