% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_test.R
\name{tds_test}
\alias{tds_test}
\title{A function to test an SVM model}
\usage{
tds_test(text_count, classification_matrix = FALSE, prop_train = 80,
  min_label = 0, max_label = 0, cost_variable = 0)
}
\arguments{
\item{text_count}{A tidy dataframe of textual data pre-processed by tds_process}

\item{classification_matrix}{A term/document matrix. Set to false (by default) to create the matrix from text_count.}

\item{prop_train}{The share of the original corpora that will be used for training the model. Default is 80\% (which means that the remaining 20\% will serve to evaluate the model).}

\item{min_label}{The minimal number of document per labels. Default is no threshold (0). If some labels are poorly represented and harder to predict it can be useful to drop them.}

\item{max_label}{The maximal number of document per labels. Default is no threshold (0). SVM is sensible to over-fitting: if a label is too represented in the original corpora, it will over-populate the results.}

\item{cost_variable}{A parametrized value for svm classification (purely optional)}
}
\value{
Raw results in a tidy format.
}
\description{
A function to test an SVM model
}
\examples{
 tds_test(text_count, prop_train = 85, min_label = 70, max_label = 150, cost_variable = 0)
}
